refactor(checks): split base.rs into module dir + dedup shell_escape
Decomposes src/checks/base.rs (986 lines) into src/checks/base/ by
responsibility. Pure code-move: behavior byte-identical.

- result.rs            CheckErrorCategory/CheckError/CheckResult + impls (184)
- context.rs           ExecutionContext enum + from_host_table (17)
- validation.rs        validate_{required,optional}_{string,bool}/file_mode (45)
- result_validation.rs StandardFields + validate_check_result_structure/
                      validate_field_naming/create_validated_result/set_*_field (167)
- execution.rs         execute_command[_with_error_handling]/is_*_error (157)
- mod.rs               submodule decls + pub use re-exports + shell_escape (17)
- tests.rs             #[cfg(test)] unit tests (400)

The three inline pub mod blocks (validation/result_validation/execution) were
already modules; their bodies moved to sibling files and the paths siblings
import (crate::checks::base::validation::*, etc.) are unchanged. CheckResult +
ExecutionContext re-exported from mod.rs; CheckError/CheckErrorCategory stay
pub in result.rs (internal-only, not re-exported).

Also dedup: shell_escape was copy-pasted byte-identically in file.rs,
package.rs, service.rs; consolidated into base/mod.rs, the three files now
import it via crate::checks::base::shell_escape.
